% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{basicsSummary}
\alias{basicsSummary}
\title{Get the very basic columns of a data frame.}
\usage{
basicsSummary(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with the \code{PP}, \code{PN}, \code{NP}, \code{NN} and
\code{AcceptedDroplets} figures.
}
\description{
It is useful to see the droplet counts for each class in addition to the
overall droplet count at a glance.
}
\examples{
df <- fullCountsSummary(KRAScountsQS)
basicsSummary(df)

}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
