% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{copiesSummary}
\alias{copiesSummary}
\title{Get the total copies per 20ul of a data frame in the context of the
basic counts.}
\usage{
copiesSummary(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with the basic data columns, mutant and wild type 
copies per 20ul, the total number of copies per 20ul, and a flag indicating 
if there are more than 2 mutant copies per 20ul.
}
\description{
Returns a data frame with the basic figures and a summary of the number of 
copies per 20ul.
}
\examples{
df <- fullCountsSummary(KRAScountsQS)
copiesSummary(df)

}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
