% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{ddpcr}
\alias{ddpcr}
\title{An environment for package variables.}
\format{An object of class \code{environment} of length 10.}
\usage{
ddpcr
}
\value{
An environment for storing package variables.
}
\description{
Stores default variables used in the package.
}
\keyword{datasets}
