% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStats.R
\name{.cov}
\alias{.cov}
\title{Get the covariance of a cluster.}
\usage{
.cov(cl, droplets, classCol)
}
\arguments{
\item{cl}{The cluster of which to find the covariance.}

\item{droplets}{A data frame of droplets with \code{Ch1.Amplitude} and
\code{Ch2.Amplitude} columns, as well as a class column (see the parameter
\code{classCol}).}

\item{classCol}{The column (name or number) from \code{droplets}
representing the class.}
}
\value{
The covariance matrix of the chosen cluster. If not defined, return
\code{NULL}.
}
\description{
Get the covariance of a single cluster.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
