% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCounts.R
\name{.dependentCols}
\alias{.dependentCols}
\title{Get a vector of all dependent columns.}
\usage{
.dependentCols(prefix = "")
}
\arguments{
\item{prefix}{A string to prepend to each of the column names. Defaults to 
\code{""} (the empty string).}
}
\value{
A vector of column names with an optional prefix.
}
\description{
Get a list of the columns in the ddPCR data that depend on the target (e.g. 
mutant or wild type). This includes columns that cannot be calculated from 
the very basic droplet counts.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
