% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCounts.R
\name{.essentialDependentCols}
\alias{.essentialDependentCols}
\title{Get a vector of essential dependent columns.}
\usage{
.essentialDependentCols(prefix = "")
}
\arguments{
\item{prefix}{A string to prepend to each of the essential column names.
Defaults to \code{""} (the empty string).}
}
\value{
A vector of essential column names with an optional prefix.
}
\description{
Get a list of the essential columns in the ddPCR data that depend on the
target (e.g. mutant or wild type).
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
