% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcrPlate.R
\name{.extractWellNames}
\alias{.extractWellNames}
\title{Retrieve the well names to use from a given list.}
\usage{
.extractWellNames(theObject, aList)
}
\arguments{
\item{theObject}{A \code{\link{ddpcrPlate}} object.}

\item{aList}{A list from which we wish to extract well names.}
}
\value{
The names of \code{aList}.
}
\description{
We simply retrieve the \code{names} from a given list, but also perform some
checks to make sure that the names are consistent with a given
\code{\link{ddpcrPlate}} object.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
