% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{.getAllSummary}
\alias{.getAllSummary}
\title{Get a summary of the number of molecules in 20ul.}
\usage{
.getAllSummary(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with mutant and wild type copies per 20ul, the total
number of copies per 20ul, and a flag indicating if there are more than
2 mutant copies per 20ul.
}
\description{
Returns a data frame with a summary of the number of mutant and wild type
copies per 20ul.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
