% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcrWell.R
\name{.getClassificationData}
\alias{.getClassificationData}
\title{Extract a classification from a data frame.}
\usage{
.getClassificationData(colName, well)
}
\arguments{
\item{colName}{The name of the column to focus on.}

\item{well}{The data frame from which to extract the classifications.}
}
\value{
A factor with levels in \code{ddpcr$classesRain}.
}
\description{
Check that non-Ch*Amplitude columns are in a known classification format and
coerce it to a factor.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
