% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{.getMutCopies}
\alias{.getMutCopies}
\title{Get the mutant copies per 20ul of a data frame.}
\usage{
.getMutCopies(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with the basic columns and mutant copies per 20ul.
}
\description{
Returns a data frame with the number of mutant copies per 20ul.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
