% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTwoDimDataFrame.R
\name{.isTwoDimDataFrame}
\alias{.isTwoDimDataFrame}
\title{Checks whether an object is a data frame with two leading double columns.}
\usage{
.isTwoDimDataFrame(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A logical value: whether the data frame meets the criteria.
}
\description{
Checks whether an object is a data frame with two leading double columns.
}
