% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahRain.R
\name{.mahDist}
\alias{.mahDist}
\title{Mahalanobis distance of droplets from a distribution.}
\usage{
.mahDist(droplets, clusStats)
}
\arguments{
\item{droplets}{A data frame of droplets with \code{Ch1.Amplitude} and
\code{Ch2.Amplitude} columns.}

\item{clusStats}{A list of statistics for a cluster generated by
\code{\link{classStats}}. This should have a \code{mean}, \code{cov} and
\code{cov.inv} values.}
}
\value{
A vector of numbers, where each one corresponds to the distance of
that droplet from the given distribution.
}
\description{
Find the Mahalanobis distance of all droplets from a given distribution.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
