% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStats.R
\name{.matrixInverse}
\alias{.matrixInverse}
\title{Get the inverse of a matrix}
\usage{
.matrixInverse(s)
}
\arguments{
\item{s}{A 2x2 matrix.}
}
\value{
The inverse of `s`, or NULL if the matrix is singular

Given a matrix, compute the inverse or return \code{NULL} if it is singular.
}
\description{
Get the inverse of a matrix
}
