% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{.renormaliseByChannel}
\alias{.renormaliseByChannel}
\title{Normalise a well on one channel only and then transform it back to the
original (combined) scale.}
\usage{
.renormaliseByChannel(wellDf, combinedCentres, wellCentres, channel)
}
\arguments{
\item{wellDf}{A data frame of the well's droplet amplitudes.}

\item{combinedCentres}{A data frame of the combined (average) centres of the
non-normalised wells.}

\item{wellCentres}{A data frame of centres corresponding to the given
\code{channel}.}

\item{channel}{An integer 1 or 2 corresponding to the channel that we are
interested in.}
}
\value{
A data frame with the rescaled amplitudes in the chosen channel.
}
\description{
Normalise a well on one channel only and then transform it back to the
original (combined) scale.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
