% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{.renormaliseWell}
\alias{.renormaliseWell}
\title{Normalise a well in both channels and then transform it back to the original
(combined) scale.}
\usage{
.renormaliseWell(allDf, well, combinedCentres, indivCentres1, indivCentres2)
}
\arguments{
\item{allDf}{A list of data frames, where each one corresponds to a well's
droplet amplitudes.}

\item{well}{The name or number of the well to normalise.}

\item{combinedCentres}{A data frame of the combined (average) centres of the
non-normalised wells.}

\item{indivCentres1}{A data frame of centres corresponding to channel 1.}

\item{indivCentres2}{A data frame of centres corresponding to channel 2.}
}
\value{
A list of data frames with the rescaled amplitudes in both channels.
}
\description{
Normalise a well in both channels and then transform it back to the original
(combined) scale.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
