% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateSummary.R
\name{.roundIt}
\alias{.roundIt}
\title{Round to at least n decimal places.}
\usage{
.roundIt(df, n = 3)
}
\arguments{
\item{df}{A data frame.}

\item{n}{How many decimal places/significant figures to round to.}
}
\value{
The data frame \code{x} with rounded entries.
}
\description{
For a data frame, if an extry is < 1, then round it to n significant
figures. If it is >= 1, then round it to n decimal places.
}
