% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcrPlate.R
\name{.slice}
\alias{.slice}
\title{Splits a long vector and according to a vector of sizes.}
\usage{
.slice(vec, wellSizes, wellNames)
}
\arguments{
\item{vec}{The vector to split.}

\item{wellSizes}{A numeric vector corresponding to sizes of the wells.}

\item{wellNames}{A character vector corresponding to the names of the wells.}
}
\value{
A list of vectors split into the given lengths.
}
\description{
Takes one long vector/factor and splits it into a list of vectors, where the
lengths are given by a vector of sizes. This may be the same as another
given list of vectors. Particularly useful if we want to combine a list of
data, do some analysis on the combined data, then split the analysis.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
