% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{.totalCopies}
\alias{.totalCopies}
\title{Get the total number of molecules in 20ul.}
\usage{
.totalCopies(df)
}
\arguments{
\item{df}{An input data frame with copies per 20ul figures.}
}
\value{
A vector listing the total numbers of copies per 20ul well.
}
\description{
Get the total number of copies of mutant and wild type molecules in a 20ul
well.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
