% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcrPlate.R
\name{elementType,SimpleList-method}
\alias{elementType,SimpleList-method}
\title{Check the types of the elements in a \code{SimpleList}.}
\usage{
\S4method{elementType}{SimpleList}(x)
}
\arguments{
\item{x}{An object.}
}
\value{
The types of objects in the \code{SimpleList} object.
}
\description{
Check the types of the elements in a \code{SimpleList}.
}
