% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{mutantCopiesSummary}
\alias{mutantCopiesSummary}
\title{Get the mutant copies per 20ul of a data frame in the context of the
basic counts.}
\usage{
mutantCopiesSummary(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with the basic columns and mutant copies per 20ul.
}
\description{
Returns a data frame with the basic figures and the number of mutant copies
per 20ul.
}
\examples{
df <- fullCountsSummary(KRAScountsQS)
mutantCopiesSummary(df)

}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
