% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisGlobal.R
\docType{data}
\name{shinyVis}
\alias{shinyVis}
\title{A Shiny environment.}
\format{
An object of class \code{environment} of length 3.
}
\usage{
shinyVis
}
\value{
An environment for storing Shiny variables.
}
\description{
Stores modes used the in Shiny UI and internal defaults.
}
\keyword{datasets}
