% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisApp.R
\name{shinyVisApp}
\alias{shinyVisApp}
\title{Shiny app.}
\usage{
shinyVisApp()
}
\value{
A Shiny application
}
\description{
A shiny app for interactive ddPCR droplet classification.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
