% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisServer.R
\name{shinyVisServer}
\alias{shinyVisServer}
\title{Shiny visualisation server.}
\usage{
shinyVisServer(input, output, session)
}
\arguments{
\item{input}{Shiny input list.}

\item{output}{Shiny output list.}

\item{session}{Shiny session.}
}
\value{
A Shiny server.
}
\description{
A Shiny server for interactive ddPCR droplet classification.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
