% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisUI.R
\name{shinyVisUI}
\alias{shinyVisUI}
\title{Shiny visualisation UI.}
\usage{
shinyVisUI()
}
\value{
A Shiny UI.
}
\description{
A Shiny UI for interactive ddPCR droplet classification.
}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
