% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVisUI.R
\name{textInputRow}
\alias{textInputRow}
\title{Inline \code{textInputs}.}
\usage{
textInputRow(inputId, label, value = "", size = NULL)
}
\arguments{
\item{inputId}{The identifier for the textInput object.}

\item{label}{The text label to display alongside.}

\item{value}{The default value of the textInput.}

\item{size}{The size (width) and maxlength.}
}
\value{
A \code{div} with the appropriate label and input field.
}
\description{
By default, \code{\link[shiny]{textInput}} objects are added to the page 
below each other. This function allows for inline \code{textInput} fields.
}
\references{
Adapted from: http://stackoverflow.com/a/21132918/1262569
}
