% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoddpcr.R
\docType{package}
\name{twoddpcr-package}
\alias{twoddpcr}
\alias{twoddpcr-package}
\title{Classifying and summarising 2-d droplet digitial PCR (ddPCR) data.}
\description{
The \code{twoddpcr} package takes droplet amplitude data from 
Bio-Rad's QuantaSoft and can classify the droplets. A summary of the 
positive/negative droplet counts can be generated, which can then be used to 
estimate the number of molecules using the Poisson distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/CRUKMI-ComputationalBiology/twoddpcr/}
  \item Report bugs at \url{http://github.com/CRUKMI-ComputationalBiology/twoddpcr/issues/}
}

}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
