% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{wildTypeCopiesSummary}
\alias{wildTypeCopiesSummary}
\title{Get the wild type copies per 20ul of a data frame in the context of the 
basic counts.}
\usage{
wildTypeCopiesSummary(df)
}
\arguments{
\item{df}{A data frame generated by \code{\link{fullCountsSummary}}.}
}
\value{
A data frame with the basic columns and wild type copies per 20ul.
}
\description{
Returns a data frame with the basic figures and the number of wild type 
copies per 20ul.
}
\examples{
df <- fullCountsSummary(KRAScountsQS)
wildTypeCopiesSummary(df)

}
\author{
Anthony Chiu, \email{anthony.chiu@cruk.manchester.ac.uk}
}
